Added support for loading scenario saves
Scenario save files are named as the scenario itself. Without this patch
only save.zip presence is checked. This is a problem because the scenario
progress is lost at every container reboot. This patch solves the issue
by adding a check for save named by FACTORIO_SCENARIO variable.

This patch also changes loading saves to pick up the desired file instead
of the latest. The latest file may not be the desired one when scenario
saves are mixed with default saves.